pre test functionality
